added step 3 and 4

+++ model/contacts.js
@@ -1,12 +1,18 @@
 const Contact = require('./schemas/contact');
 
-const listContacts = async () => {
-  const result = await Contact.find({}, { __v: 0 })
+const listContacts = async (userId) => {
+  const result = await Contact.find({ owner: userId }).populate({
+    path: 'owner',
+    select: 'email subscription -_id',
+  });
   return result;
 };
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findOne({ _id: contactId }).select("-__v");
+const getContactById = async (contactId, userId) => {
+  const result = await Contact.findOne({ _id: contactId, owner: userId, }).populate({
+    path: 'owner',
+    select: 'email subscription -_id',
+  });
   return result;
 };
 
@@ -15,17 +21,17 @@ const addContact = async (body) => {
   return result;
 };
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, userId) => {
   const result = await Contact.findByIdAndUpdate(
-    contactId,
+    { contactId, owner: userId },
     ...body,
     { new: true },
   );
   return result;
 };
 
-const removeContact = async (contactId) => {
-  const result = await Contact.findByIdAndRemove(contactId);
+const removeContact = async (contactId, userId) => {
+  const result = await Contact.findByIdAndRemove({ contactId, owner: userId });
   return result;
 };
